requirements.txt
asgiref==3.2.10
autopep8==1.6.0
certifi==2021.5.30
charset-normalizer==2.0.3
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
django-hosts==4.0
django-jet-reboot==1.1.1
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
google-api-python-client==1.4.1
gunicorn==20.1.0
httplib2==0.20.2
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauth2client==4.1.3
packaging==21.0
phonenumbers==8.12.28
Pillow==8.3.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==2.4.7
python-decouple==3.5
python-magic==0.4.24
pytz==2021.1
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
